+++ kaye/prompt/prompt_blueprint.py
@@ -107,9 +107,7 @@ class PromptBlueprint(dict):
                 parent = prev_node.ancestors[level - 1]
 
             # create/add node
-            node = bp._parse_add_dynamic_node(
-                heading, parent
-            ) or bp._parse_add_corpus_node(parent, heading, line)
+            node = bp._parse_add_corpus_node(parent, heading, line)
 
             # include node in the blueprint
             bp[hash(node)] = is_checkmarked
@@ -182,7 +180,7 @@ class PromptBlueprint(dict):
     # node operations  *********************************************************
     def is_checkmarked(self, node):
         """
-        :param key: node object; hash value; name or identifier
+        :param key: node object; hash value; name
         :type node: BasePromptNode or int or str
         :raises TypeError:
         :raise ValueError:
@@ -199,7 +197,7 @@ class PromptBlueprint(dict):
         (will add node into this blueprint if not, then checkmarked it)
 
 
-        :param key: node object; hash value; name or identifier
+        :param key: node object; hash value; name
         :type node: BasePromptNode or int or str
         :param recursively: allow checkmarks on node's descendants,
                 defaults to False
@@ -217,7 +215,7 @@ class PromptBlueprint(dict):
         (node must be contained in this blueprint)
 
 
-        :param key: node object; hash value; name or identifier
+        :param key: node object; hash value; name
         :type node: BasePromptNode or int or str
         :param recursively: allow checkmarks on node's descendants,
                 defaults to False
@@ -282,7 +280,7 @@ class PromptBlueprint(dict):
                 checkmark_prefix = EMPTY_PREFIX
 
             # e.g. "[x] │   └── Capitalization Style"
-            node_line = checkmark_prefix + pre + node.identifier
+            node_line = checkmark_prefix + pre + node.name
             lines.append(node_line)
 
             # lines for content preview part
@@ -414,26 +412,6 @@ class PromptBlueprint(dict):
 
         return "{}Kaye v{}".format(name_part, kaye_version)
 
-    def _parse_add_dynamic_node(self, heading, parent):
-        # Bug adds multiple nodes
-        # early exit for non-dynamic node
-        if not DynamicNode.ID_PATTERN.match(heading):
-            return False
-
-        name = heading[1:-1]
-
-        # decide type of dynamic node by name's pattern
-        if name == TodayNode.HEADING:
-            return TodayNode(parent)
-        elif name == AbbrNode.HEADING:
-            return AbbrNode(parent)
-        elif name == PLCNode.HEADING:
-            return PLCNode(parent)
-        elif name == UsableAbbrNode.HEADING:
-            return UsableAbbrNode(parent)
-        else:
-            return False
-
     def _parse_add_corpus_node(self, parent, heading, line):
         """
         find current node when non-dynamic, and include it in this blueprint
@@ -512,7 +490,7 @@ class PromptBlueprint(dict):
         - ``.is_checkmarked()``
         - ``.__contains__()``
 
-        to search a node in corpus providing node object/name/identifier/hash
+        to search a node in corpus providing node object/name/hash
 
 
         :raises TypeError:
@@ -520,21 +498,19 @@ class PromptBlueprint(dict):
         """
         corpus_and_descendants = [self.corpus] + list(self.corpus.descendants)
 
-        # search by name/identifier  -------------------------------------------
+        # search by name   -----------------------------------------------------
         if isinstance(node_arg, str):
             node_obj = None
 
-            # search all descendants with name/identifier of node
+            # search all descendants with name of node
             for n in corpus_and_descendants:
-                if node_arg in (n.name, n.identifier):
+                if node_arg == n.name:
                     node_obj = n
                     break
 
             if node_obj is None:
                 raise ValueError(
-                    "no node in corpus with name/identifier: {}".format(
-                        repr(node_arg)
-                    )
+                    "no node in corpus with name: {}".format(repr(node_arg))
                 )
 
             node_hash = hash(node_obj)
@@ -568,8 +544,9 @@ class PromptBlueprint(dict):
 
         else:
             raise TypeError(
-                "must be BasePromptNode/"
-                "int(hash value)/str(name/identifier): {}".format(node_arg)
+                "must be BasePromptNode/int(hash value)/str(name): {}".format(
+                    node_arg
+                )
             )
 
         return node_obj, node_hash
@@ -581,7 +558,7 @@ class PromptBlueprint(dict):
         allow ``PromptBlueprint`` to perform membership tests
 
 
-        :param key: node object; hash value; name or identifier
+        :param key: node object; hash value; name
         :type key: PromptCorpusNode or int or str
         :raises ValueError:
         :return: if blueprint contains the node