+++ tests/prompt/base/__init__.py
@@ -3,9 +3,5 @@ from kaye.prompt.base_prompt_node import BasePromptNode
 
 class UnitTestNode(BasePromptNode):
 
-    @property
-    def identifier(self):
-        return self.name.upper()
-
     def content_lines(self, *args, **kwargs):
         return self.lines