+++ tests/prompt/bp/prompt-bp-checkmark_test.py
@@ -175,11 +175,7 @@ class Test11:  # ===============================================================
         opt = exec_info.value.args[0]
         print(opt)
 
-        assert (
-            opt
-            == "must be BasePromptNode/"
-            "int(hash value)/str(name/identifier): 12.5"
-        )
+        assert opt == "must be BasePromptNode/int(hash value)/str(name): 12.5"
 
     def test_bad_type2(_, local_testee11):
         bp, _ = local_testee11
@@ -194,7 +190,7 @@ class Test11:  # ===============================================================
         assert (
             opt
             == "must be BasePromptNode/"
-            "int(hash value)/str(name/identifier): ['a', 'b', 'c']"
+            "int(hash value)/str(name): ['a', 'b', 'c']"
         )
 
     def test_bad_str_no_found1(_, local_testee11):
@@ -207,7 +203,7 @@ class Test11:  # ===============================================================
         opt = exec_info.value.args[0]
         print(opt)
 
-        assert opt == "no node in corpus with name/identifier: 'AAAZZZ'"
+        assert opt == "no node in corpus with name: 'AAAZZZ'"
 
     def test_bad_hash(self, local_testee11):
         bp, _ = local_testee11
@@ -462,9 +458,7 @@ class TestDynamicNodes:
     def test_checkmark_by_obj1(_, local_dynamic_testee1):
         bp = local_dynamic_testee1
 
-        node = bp.corpus["Main Title"]["Introduction"]["Background"][
-            "Importance"
-        ]["Abbreviations"]
+        node = bp.corpus["{Abbreviations}"]
 
         assert bp.checkmark(node)
         assert bp.is_checkmarked(node)
@@ -472,9 +466,7 @@ class TestDynamicNodes:
     def test_checkmark_by_hash1(_, local_dynamic_testee1):
         bp = local_dynamic_testee1
 
-        node = bp.corpus["Main Title"]["Introduction"]["Background"][
-            "Importance"
-        ]["Abbreviations"]
+        node = bp.corpus["{Abbreviations}"]
         node_hash = hash(node)
 
         assert bp.checkmark(node_hash)
@@ -483,10 +475,8 @@ class TestDynamicNodes:
 
     def test_checkmark_by_name1(_, local_dynamic_testee1):
         bp = local_dynamic_testee1
-        node = bp.corpus["Main Title"]["Introduction"]["Background"][
-            "Importance"
-        ]["Abbreviations"]
-        ipt = "Abbreviations"
+        node = bp.corpus["{Abbreviations}"]
+        ipt = "{Abbreviations}"
 
         assert bp.checkmark(ipt)
 
@@ -494,9 +484,7 @@ class TestDynamicNodes:
 
     def test_checkmark_by_identifier1(_, local_dynamic_testee1):
         bp = local_dynamic_testee1
-        node = bp.corpus["Main Title"]["Introduction"]["Background"][
-            "Importance"
-        ]["Abbreviations"]
+        node = bp.corpus["{Abbreviations}"]
         ipt = "{Abbreviations}"
 
         assert bp.checkmark(ipt)
@@ -508,7 +496,7 @@ class TestDynamicNodes:
     def test_checkmark_by_obj2(_, local_dynamic_testee1):
         bp = local_dynamic_testee1
 
-        node = bp.corpus["Main Title"]["Methods"]["Programming Languages Code"]
+        node = bp.corpus["{Programming Languages Code}"]
 
         assert bp.checkmark(node)
         assert bp.is_checkmarked(node)
@@ -516,7 +504,7 @@ class TestDynamicNodes:
     def test_checkmark_by_hash2(_, local_dynamic_testee1):
         bp = local_dynamic_testee1
 
-        node = bp.corpus["Main Title"]["Methods"]["Programming Languages Code"]
+        node = bp.corpus["{Programming Languages Code}"]
         node_hash = hash(node)
 
         assert bp.checkmark(node_hash)
@@ -526,8 +514,8 @@ class TestDynamicNodes:
     def test_checkmark_by_name2(_, local_dynamic_testee1):
         bp = local_dynamic_testee1
 
-        node = bp.corpus["Main Title"]["Methods"]["Programming Languages Code"]
-        ipt = "Programming Languages Code"
+        node = bp.corpus["{Programming Languages Code}"]
+        ipt = "{Programming Languages Code}"
 
         assert bp.checkmark(ipt)
 
@@ -535,7 +523,7 @@ class TestDynamicNodes:
 
     def test_checkmark_by_identifier2(_, local_dynamic_testee1):
         bp = local_dynamic_testee1
-        node = bp.corpus["Main Title"]["Methods"]["Programming Languages Code"]
+        node = bp.corpus["{Programming Languages Code}"]
         ipt = "{Programming Languages Code}"
 
         bp.checkmark(ipt)
@@ -547,20 +535,21 @@ class TestDynamicNodes:
     def test_recursive1(_, dynamic_bp_testee_empty):
         bp = dynamic_bp_testee_empty
         answer = """    ○
-[x] └── Main Title
-[x]     ├── Introduction
-[x]     │   ├── Background
-[x]     │   │   └── Importance
-[x]     │   │       ├── Objective
-[x]     │   │       └── {Abbreviations}
-[x]     │   └── {Usable Abbreviations}
-[x]     ├── Methods
-[x]     │   ├── Data Collection
-[x]     │   │   └── Tools Used
-[x]     │   │       └── Future Work
-[x]     │   │           └── {Today}
-[x]     │   └── {Programming Languages Code}
-[x]     └── Conclusion"""
+[x] ├── Main Title
+[x] │   ├── Introduction
+[x] │   │   └── Background
+[x] │   │       └── Importance
+[x] │   │           └── Objective
+[x] │   ├── Methods
+[x] │   │   └── Data Collection
+[x] │   │       └── Tools Used
+[x] │   │           └── Future Work
+[x] │   └── Conclusion
+[ ] ├── {Today}
+[ ] ├── {Abbreviations}
+[ ] ├── {Usable Abbreviations}
+[ ] ├── {Languages Code}
+[ ] └── {Programming Languages Code}"""
 
         bp.checkmark("Main Title", recursively=True)
 
@@ -572,21 +561,23 @@ class TestDynamicNodes:
 
     def test_recursive2(_, dynamic_bp_testee_empty):
         bp = dynamic_bp_testee_empty
+
         answer = """    ○
-[ ] └── Main Title
-[x]     ├── Introduction
-[x]     │   ├── Background
-[x]     │   │   └── Importance
-[x]     │   │       ├── Objective
-[x]     │   │       └── {Abbreviations}
-[x]     │   └── {Usable Abbreviations}
-[ ]     ├── Methods
-[ ]     │   ├── Data Collection
-[ ]     │   │   └── Tools Used
-[ ]     │   │       └── Future Work
-[ ]     │   │           └── {Today}
-[ ]     │   └── {Programming Languages Code}
-[ ]     └── Conclusion"""
+[ ] ├── Main Title
+[x] │   ├── Introduction
+[x] │   │   └── Background
+[x] │   │       └── Importance
+[x] │   │           └── Objective
+[ ] │   ├── Methods
+[ ] │   │   └── Data Collection
+[ ] │   │       └── Tools Used
+[ ] │   │           └── Future Work
+[ ] │   └── Conclusion
+[ ] ├── {Today}
+[ ] ├── {Abbreviations}
+[ ] ├── {Usable Abbreviations}
+[ ] ├── {Languages Code}
+[ ] └── {Programming Languages Code}"""
 
         bp.checkmark("Introduction", recursively=True)