+++ tests/prompt/dn/prompt-dn-dynamic_node_test.py
@@ -12,6 +12,10 @@ from kaye.prompt.base_prompt_node import DynamicNode
 # helpers  #####################################################################
 class DynamicTestee(DynamicNode):
 
+    # implement DynamicNode  ===================================================
+
+    HEADING = "Testee Node"
+
     # implement BasePromptNode  ================================================
 
     def content_lines(self, **kwargs):
@@ -32,33 +36,34 @@ class DynamicTestee(DynamicNode):
 
 @pytest.fixture
 def testee1():
-    return DynamicTestee("Testee Node", None)
+    return DynamicTestee(None)
 
 
-class TestId:  #################################################################
+class TestName:  #################################################################
 
     def test1(_, testee1):
-        print(testee1.identifier)
-        assert testee1.identifier == "{Testee Node}"
+        opt = testee1.name
+        print(opt)
+        assert opt == "{Testee Node}"
 
 
 class TestNoChild:  ############################################################
 
     def test1(_, testee1):
         with pytest.raises(TypeError) as exec_info:
-            DynamicTestee("Parent", None, children=[testee1])
+            DynamicTestee(None, children=[testee1])
 
         opt = exec_info.value.args[0]
         print(opt)
         assert (
             opt
-            == "<class 'tests.prompt.prompt-dynamic_node_test.DynamicTestee'> "
+            == "<class 'prompt-dn-dynamic_node_test.DynamicTestee'> "
             "must be leaf node"
         )
 
     def test2(_, testee1):
         with pytest.raises(TypeError) as exec_info:
-            parent = DynamicTestee("Parent", None)
+            parent = DynamicTestee(None)
             parent.children = [testee1]
 
         opt = exec_info.value.args[0]
@@ -66,6 +71,6 @@ class TestNoChild:  ############################################################
 
         assert (
             opt
-            == "<class 'tests.prompt.prompt-dynamic_node_test.DynamicTestee'> "
+            == "<class 'prompt-dn-dynamic_node_test.DynamicTestee'> "
             "must be leaf node"
         )