+++ tests/prompt/dn/prompt-dn-plc_test.py
@@ -30,8 +30,7 @@ class TestInit:  ###############################################################
 
     def test1(_, testee1, local_corpus_testee1):
         assert testee1.parent is local_corpus_testee1
-        assert testee1.name == "Programming Languages Code"
-        assert testee1.identifier == "{Programming Languages Code}"
+        assert testee1.name == "{Programming Languages Code}"
 
     def test_preview1(_, local_corpus_testee1):
         opt = local_corpus_testee1.generate_prompt_tree_preview(
@@ -43,7 +42,7 @@ class TestInit:  ###############################################################
 │   ├── Description
 │   ├── Installation
 │   └── License
-└── Programming Languages Code"""
+└── {Programming Languages Code}"""
 
 
 class TestCopy:  ###############################################################
@@ -52,14 +51,14 @@ class TestCopy:  ###############################################################
         copied = copy.copy(testee1)
 
         assert isinstance(copied, PLCNode)
-        assert copied.name == "Programming Languages Code"
+        assert copied.name == "{Programming Languages Code}"
         assert copied.parent is None
 
     def test_deep_copy1(_, testee1):
         copied = copy.deepcopy(testee1)
 
         assert isinstance(copied, PLCNode)
-        assert copied.name == "Programming Languages Code"
+        assert copied.name == "{Programming Languages Code}"
         assert copied.parent is None