+++ tests/prompt/dn/prompt-dn-usable_test.py
@@ -1,9 +1,9 @@
 """
-prompt-lc_test.py
+prompt-usable_test.py
 
 Unit Tests (using pytest) for:
 
-LanguageCodeNode
+- UsableAbbrNode
 """
 
 import copy
@@ -12,7 +12,7 @@ import copy
 import pytest
 
 
-from kaye.prompt.abbr_nodes import LanguageCodeNode
+from kaye.prompt.abbr_nodes import UsableAbbrNode
 
 
 # pytest fixtures  #############################################################
@@ -23,18 +23,14 @@ def local_corpus_testee1(corpus_testee1):
 
 @pytest.fixture(scope="session")
 def testee1(local_corpus_testee1):
-    return LanguageCodeNode(local_corpus_testee1)
+    return UsableAbbrNode(local_corpus_testee1)
 
 
-# pytest  ######################################################################
-
-
-class TestInit:  # =============================================================
+class TestInit:  ###############################################################
 
     def test1(_, testee1, local_corpus_testee1):
         assert testee1.parent is local_corpus_testee1
-        assert testee1.name == "Languages Code"
-        assert testee1.identifier == "{Languages Code}"
+        assert testee1.name == "{Usable Abbreviations}"
 
     def test_preview1(_, local_corpus_testee1):
         opt = local_corpus_testee1.generate_prompt_tree_preview(
@@ -46,35 +42,29 @@ class TestInit:  # =============================================================
 │   ├── Description
 │   ├── Installation
 │   └── License
-└── Languages Code"""
+└── {Usable Abbreviations}"""
 
 
-class TestCopy:  # =============================================================
+class TestCopy:  ###############################################################
 
     def test_copy1(_, testee1):
         copied = copy.copy(testee1)
 
-        assert isinstance(copied, LanguageCodeNode)
-        assert copied.name == "Languages Code"
+        assert isinstance(copied, UsableAbbrNode)
+        assert copied.name == "{Usable Abbreviations}"
         assert copied.parent is None
 
     def test_deep_copy1(_, testee1):
         copied = copy.deepcopy(testee1)
 
-        assert isinstance(copied, LanguageCodeNode)
-        assert copied.name == "Languages Code"
+        assert isinstance(copied, UsableAbbrNode)
+        assert copied.name == "{Usable Abbreviations}"
         assert copied.parent is None
 
 
-class TestContentLines:  # =====================================================
+class TestContentLines:  #######################################################
 
     def test1(_, testee1):
         opt = testee1.content_lines()
         print(opt)
-        assert opt == [
-            "-`de`:Deutsch",
-            "-`en`:English",
-            "-`zh`:中文",
-            "-`zhs`:大陆简体中文",
-            "-`zht`:香港繁體中文",
-        ]
+        assert opt == ["-`&`:and", "-`/`:or"]